requirements.txt
sseclient==0.0.26
GitPython==3.1.1
requests==2.21.0